Selected meals and prices working

+++ DTO/CheckBoxMeals.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string LabelName { get; set; }
         public float Price { get; set; }
-        public int CountOfMeals { get; set; }
+        public float CountOfMeals { get; set; }
         public bool IsChecked { get; set; }
 
     }
